requirements.txt
braintree==3.19.0
Django==1.8.4
django-cors-headers==1.1.0
django-crispy-forms==1.5.0
django-filter==0.11.0
django-registration-redux==1.2
djangorestframework==3.3.3
Pillow==2.9.0
pkg-resources==0.0.0
requests==2.7.0
